More library files
This brings the archive into synch with my harddisk...

Everything in the lib directory should compile without problems (yes, even
without warnings even if WARN=1). It is not well-tested, though, and there
are some minor functions which must still be added. It is now almost
usable.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@95 7894878c-1315-0410-8ee3-d5d059ff63e0

+++ lib/chips.c
@@ -35,7 +35,7 @@
    divided. */
 
 
-#define VALUE(x) (x-1) * sizeof(double)
+#define VALUE(x) (x-1) * sizeof(long)
 
 static sensors_chip_feature lm78_features[] =
   {
